Cap7 - Drawing in Direct3D II - LandAndWaves

+++ Common/Source/D3DApp.cpp
@@ -245,7 +245,7 @@ LRESULT D3DApp::MsgProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam)
 		return 0;
 
 	case WM_KEYUP:
-		if ((int)wParam == VK_ESCAPE)
+		if (wParam == VK_ESCAPE)
 		{
 			PostQuitMessage(0);
 		}
